+++ lib/web_ui/lib/src/engine/compositor/canvas_kit_canvas.dart
@@ -79,7 +79,7 @@ class CanvasKitCanvas implements ui.Canvas {
   }
 
   @override
-  void transform(Float64List matrix4) {
+  void transform(Float32List matrix4) {
     assert(matrix4 != null);
     if (matrix4.length != 16) {
       throw ArgumentError('"matrix4" must have 16 entries.');
@@ -87,7 +87,7 @@ class CanvasKitCanvas implements ui.Canvas {
     _transform(matrix4);
   }
 
-  void _transform(Float64List matrix4) {
+  void _transform(Float32List matrix4) {
     _canvas.transform(matrix4);
   }
