+++ lib/web_ui/lib/src/ui/window.dart
@@ -1108,7 +1108,7 @@ class PluginUtilities {
 
 // TODO(flutter_web): see https://github.com/flutter/flutter/issues/33616.
 class ImageShader implements Shader {
-  ImageShader(Image image, TileMode tmx, TileMode tmy, Float64List matrix4);
+  ImageShader(Image image, TileMode tmx, TileMode tmy, Float32List matrix4);
 }
 
 // TODO(flutter_web): probably dont implement this one.
